+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changes in first release version of MCEq from release candidate 1:
 
+Current:
+- multiple calls to `set_single_particle` can define an initial state by using an the `append` flag
+
 Version 1.0.9:
 - disable_decays flag in advanced options fixed
 - threshold energy not used in n_mu, n_e